+++ src/client/Client.ts
@@ -537,20 +537,14 @@ class Client extends EventEmitter {
 	 * Sends a request to the server to join a room with specific name.
 	 */
 	enterRoom(name: string, options?: RoomJoinOptions) {
-		if (options?.password) {
-			this.main.send(
-				BulleIdentifier.roomPassworded,
-				new ByteArray().writeUTF(options.password).writeUTF(name)
-			);
-		} else {
-			this.main.send(
-				BulleIdentifier.room,
-				new ByteArray()
-					.writeUTF(options?.language ?? this.language)
-					.writeUTF(name)
-					.writeBoolean(options?.auto ?? false)
-			);
-		}
+		this.main.send(
+			BulleIdentifier.room,
+			new ByteArray()
+				.writeUTF(options?.language ?? this.language)
+				.writeUTF(name)
+				.writeUTF(options?.password ?? "")
+				.writeBoolean(options?.auto ?? false)
+		);
 	}
 
 	/**