+++ CHANGELOG.md
@@ -8,9 +8,10 @@ It is recommended to keep the client up-to-date with latest versions to avoid po
 This release includes some important fixes to run on the latest protocol.
 
 ### Added
-- `Client`: new event `roomPassworded` is emitted when a room being attempted to enter needs a password.
+- `Client`: New event `roomPassworded` is emitted when a room being attempted to enter needs a password.
 
 ### Changes
+- `BREAKING` `Client:` Event `disconnect` will also be emitted when a connection error occurs. It will also include an optional `err` parameter, which will be `null` when the disconnection is client-triggered. 
 - Support server updates to room joining. `Client.enterRoom` works again. `BulleIdentifier.roomPassworded` (5,39) is no longer server-bound packet and is replaced by having the password in the `BulleIdentifier.room` (5,38) packet.
 
 ### Fixed