+++ CHANGELOG.md
@@ -6,6 +6,16 @@ It is recommended to keep the client up-to-date with latest versions to avoid po
 ## 1.2.0-beta.1 - Unreleased
 
 ### Changes
+- Export `Connection` class.
+- Export `TFMConnectionError` class. Added `TFMConnectionError.isThis(payload: any)` static method to narrow error types.
+```ts
+const castedConnErr: any = new TFMConnectionError("bulle") as any
+
+if (TFMConnectionError.isThis(castedConnErr)) {
+	// castedConnErr in this scope is narrowed from `any` into `TFMConnectionError`
+	console.log(castedConnErr.serverType); // prints "bulle"
+}
+```
 -->
 
 ## 1.1.0 - 2023-06-03
@@ -16,11 +26,11 @@ This release includes some important fixes to run on the latest protocol.
 
 ### Changes
 - `BREAKING` `Client:` Event `disconnect` will also be emitted when a connection error occurs. It will also include an optional `err` parameter, which will be `null` when the disconnection is client-triggered. ([#4](https://github.com/cheeseformice/transformice.js/issues/4), @Cassolette)
-- Support server updates to room joining. `Client.enterRoom` works again. `BulleIdentifier.roomPassworded` (5,39) is no longer server-bound packet and is replaced by having the password in the `BulleIdentifier.room` (5,38) packet.
 
 ### Fixed
 - Cheeseformice endpoint is dead. Replaced it with a working link and now `Client.fetchIP` works again. 
 - `Client.loadLua` actually works. It previously did not include the script buffer in its outgoing packets.
+- Support server updates to room joining. `Client.enterRoom` works again. `BulleIdentifier.roomPassworded` (5,39) is no longer server-bound packet and is replaced by having the password in the `BulleIdentifier.room` (5,38) packet.
 
 ## 1.0.1 - 2022-07-07
 This release includes some important fixes to run on the latest protocol.